activeMQ/python-client/requirements.txt
stomp.py>=8.0.0
